apply: Use batch-source action plans
Apply-from builds deferred plans for text, binary, and submodule pointer
actions before opening the undo checkpoint. Include-from has the same
binary and submodule plan shapes.

The command lacks a support boundary for resource-owning batch-source
action plans, so apply-from defines binary and submodule records beside its
command flow. That makes the shared plan lifetime rule command-local.

This commit begins separating batch-source action planning by adding
commands.batch_source.action_plans and moving apply-from's binary and
submodule plans plus close loop onto it.

Subsequent commits will guard the apply adopter and move include-from onto
the shared plan records.
