+++ src/git_stage_batch/commands/batch_source/__init__.py
@@ -1,0 +1,1 @@
+"""Support helpers for commands that consume batch sources."""
